Big update

+++ src/main/java/com/example/ws/microservices/firstmicroservices/dto/templateTables/DepartmentDTO.java
@@ -11,7 +11,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 public class DepartmentDTO implements SiteAware{
-    private Short id;
+
+    private Integer id;
     private String name;
     private String siteName;
